requirements.txt
asgiref==3.7.2
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==22.1
djangorestframework==3.13.1
factory-boy==3.0.1
Faker==19.1.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
